+++ src/main/java/com/codebrew/repository/EventsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventsRepository extends JpaRepository<Events, Long> {
-    Events findEventById(Long id);
+public interface EventsRepository extends JpaRepository<Events, Integer> {
+    Events findEventById(Integer id);
 
 }
